+++ Src-graphs/SVGrafsM_Im.cpp
@@ -434,7 +434,7 @@ void SVGrafsM_Im::updateRHSFlags()
 
     const QBitArray &saveBits = p.im.prbj[ip].sns.saveBits;
 
-    for( int ic = 0, nC = ic2Y.size(); ic < nC; ++ic ) {
+    for( int ic = 0, nC = (int)ic2Y.size(); ic < nC; ++ic ) {
 
         MGraphY &Y = ic2Y[ic];
