+++ Src-graphs/SVGrafsM_Ob.cpp
@@ -285,7 +285,7 @@ void SVGrafsM_Ob::updateRHSFlags()
 
     const QBitArray &saveBits = p.im.get_iStrOneBox( ip ).sns.saveBits;
 
-    for( int ic = 0, nC = ic2Y.size(); ic < nC; ++ic ) {
+    for( int ic = 0, nC = (int)ic2Y.size(); ic < nC; ++ic ) {
 
         MGraphY &Y = ic2Y[ic];
